- Further implemented authentication with auth0

+++ wa-quizolai-business/src/dao/question.dao.ts
@@ -1,21 +1,35 @@
 import { Question } from '../models/question.model';
 import { IQuestion, IInputQuestion } from "../../../wa-quizolai-shared";
 import { StatusCodeError } from '../util/exceptions/exceptions';
+import {hasPermission} from "../service/auth.service";
 
 const questionDao = {
-    async saveQuestion(inputQuestion: IInputQuestion): Promise<IQuestion> {
-        const question = new Question(inputQuestion);
-        return await question.save();
+    async saveQuestion(inputQuestion: IInputQuestion, userID : string): Promise<IQuestion> {
+        const nq = new Question(inputQuestion);
+        nq.userID = userID;
+        return await nq.save();
     },
 
-    async saveQuestions(inputQuestions: IInputQuestion[]): Promise<IQuestion[]> {
-        const questions = inputQuestions.map(inputQuestion => new Question(inputQuestion));
-        console.log(questions)
+    async saveQuestions(inputQuestions: IInputQuestion[], userID : string): Promise<IQuestion[]> {
+        const questions = inputQuestions.map(inputQuestion => {
+           const nq =  new Question(inputQuestion);
+           nq.userID = userID;
+           return nq;
+        });
         return await Question.insertMany(questions);
     },
 
-    async deleteQuestion(id: string) {
-        const question = await Question.findOneAndDelete({ _id: id });
+    async deleteQuestion(id: string, userID : string, permissions : string[]): Promise<IQuestion> {
+
+        let question = await Question.findOne({ _id: id });
+        if (!question) {
+            throw new StatusCodeError( 'Question not found', 404);
+        }
+        if (question.userID !== userID && !hasPermission(["edit:question"], permissions)) {
+            throw new StatusCodeError(`You don't have permission to delete this question`, 403);
+        }
+        question = await Question.findOneAndDelete({ _id: id });
+
         if (!question) {
             throw new StatusCodeError(`Question with id ${id} not found`, 404);
         }
@@ -30,7 +44,16 @@ const questionDao = {
         return question;
     },
 
-    async updateQuestion(id: string, inputQuestion: IInputQuestion): Promise<IQuestion> {
+    async updateQuestion(id: string, inputQuestion: IInputQuestion, userID : string, permissions : string[]): Promise<IQuestion> {
+        const question = await Question.findOne({ _id: id });
+
+        if (!question) {
+            throw new StatusCodeError(`Question with id ${id} not found`, 404);
+        }
+        if (question.userID !== userID && !hasPermission(["edit:question"], permissions)) {
+            throw new StatusCodeError(`You don't have permission to edit this question`, 403);
+        }
+
         const updatedQuestion = await Question.findOneAndUpdate({ _id: id }, inputQuestion, { new: true, upsert: false });
         if (!updatedQuestion) {
             throw new StatusCodeError(`Question with id ${id} not found`, 404);
